+++ weave/init_message.py
@@ -1,6 +1,6 @@
 import typing
 
-from weave import urls
+from weave.legacy import urls
 
 if typing.TYPE_CHECKING:
     import packaging.version  # type: ignore[import-not-found]
