+++ weave/trace/weave_client.py
@@ -9,8 +9,9 @@ from typing import Any, Dict, Iterator, Optional, Sequence, Union
 import pydantic
 from requests import HTTPError
 
-from weave import urls, version
+from weave import version
 from weave.client_context import weave_client as weave_client_context
+from weave.legacy import urls
 from weave.trace import call_context, trace_sentry
 from weave.trace.exception import exception_to_json_str
 from weave.trace.feedback import FeedbackQuery, RefFeedbackQuery
